pyproject.toml
[project]
name = "ai-coding-fe"
version = "0.1.0"
description = ""
authors = [
    {name = "Carlos Ribeiro",email = "carlosmlribeiro@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.46.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
